Lesson16_facade_add_base_fasade/requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
async-generator==1.10
attrs==23.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
enum34==1.1.10
exceptiongroup==1.1.1
flake8==6.0.0
h11==0.14.0
idna==3.4
iniconfig==2.0.0
lxml==4.9.2
mccabe==0.7.0
namedlist==1.8
outcome==1.2.0
packaging==23.0
pluggy==1.0.0
psycopg2-binary==2.9.6
py==1.11.0
pycodestyle==2.10.0
pyflakes==3.0.1
pyjava==0.4.0
PySocks==1.7.1
pytest==7.3.0
pytest-html==3.2.0
pytest-metadata==2.0.4
python-dotenv==1.0.0
requests==2.30.0
selenium==4.8.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
urllib3==1.26.15
webdriver-manager==3.8.6
wsproto==1.2.0
